# openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: 'http://localhost'
    description: Generated server url
paths:
...
  '/api/flows/{id}':
    get:
      tags:
        - flow-controller
      summary: Получить поток по ID
      operationId: getFlow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешно
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FlowResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Поток не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
    put:
      tags:
        - flow-controller
      summary: Обновить поток
      operationId: updateFlow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlowRequest'
        required: true
      responses:
        '200':
          description: Поток обновлён
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FlowResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Нет доступа к команде
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Поток не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
    delete:
      tags:
        - flow-controller
      summary: Удалить поток
      operationId: deleteFlow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Поток удалён
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Поток не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
...
  /api/flows:
    get:
      tags:
        - flow-controller
      summary: Получить список потоков по команде
      operationId: listFlows
      parameters:
        - name: teamId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Список потоков
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlowResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Нет доступа к команде
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
    post:
      tags:
        - flow-controller
      summary: Создание потока
      operationId: createFlow
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlowRequest'
        required: true
      responses:
        '200':
          description: Поток создан
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FlowResponse'
        '400':
          description: Ошибка валидации или аргумента
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Пользователь не состоит в команде
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    AIFilter:
      type: object
      properties:
        trigger:
          type: string
        prompt:
          type: string
    AddPhotoFilter:
      type: object
      properties:
        trigger:
          type: string
        photoUrl:
          type: string
    Filter:
      type: object
      properties:
        addPhoto:
          $ref: '#/components/schemas/AddPhotoFilter'
        removePhoto:
          $ref: '#/components/schemas/RemovePhotoFilter'
        textReplace:
          $ref: '#/components/schemas/TextReplaceFilter'
        aiFilter:
          $ref: '#/components/schemas/AIFilter'
    FlowRequest:
      type: object
      properties:
        teamId:
          type: string
          format: uuid
        source:
          $ref: '#/components/schemas/Source'
        sinks:
          type: array
          items:
            $ref: '#/components/schemas/Sink'
          maxItems: 2147483647
          minItems: 1
        requiresModeration:
          type: boolean
        filters:
          type: array
          items:
            $ref: '#/components/schemas/Filter'
      required:
        - filters
        - requiresModeration
        - sinks
        - source
        - teamId
    RemovePhotoFilter:
      type: object
      properties:
        trigger:
          type: string
    Sink:
      type: object
      properties:
        telegramSink:
          $ref: '#/components/schemas/TelegramSink'
    Source:
      type: object
      properties:
        telegramSource:
          $ref: '#/components/schemas/TelegramSource'
    TelegramSink:
      type: object
      properties:
        chatId:
          type: string
    TelegramSource:
      type: object
      properties:
        chatId:
          type: string
    TextReplaceFilter:
      type: object
      properties:
        trigger:
          type: string
        pattern:
          type: string
        replacement:
          type: string
    FlowResponse:
      type: object
      properties:
        id:
          type: string
        teamId:
          type: string
          format: uuid
        source:
          $ref: '#/components/schemas/Source'
        sinks:
          type: array
          items:
            $ref: '#/components/schemas/Sink'
        filters:
          type: array
          items:
            $ref: '#/components/schemas/Filter'
        requiresModeration:
          type: boolean
        updatedAt:
          type: string
          format: date-time
...
    StatusResponse:
      type: object
      properties:
        status:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT